+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+# Changelog
+
 ## Unreleased
 
-- ???
+### Fixed
+
+- rollover writes local date instead of utc
 
 ## v0.0.2 - 2025-05-24